+++ test/test_corpus.py
@@ -68,19 +68,27 @@ CORPUS_SETTINGS = {
 
 
 @pytest.mark.parametrize("schema", get_names(schemas), indirect=["schema"])
-@settings(**CORPUS_SETTINGS)
-@given(data=st.data())
-def test_corpus(data, schema: Schema, validate_operation):
-    query = data.draw(from_schema(schema.raw, custom_scalars=schema.custom_scalars))
-    validate_operation(schema.raw, query)
+def test_corpus(schema: Schema, validate_operation):
+    strategy = from_schema(schema.raw, custom_scalars=schema.custom_scalars)
+
+    @settings(**CORPUS_SETTINGS)
+    @given(query=strategy)
+    def inner(query):
+        validate_operation(schema.raw, query)
+
+    inner()
 
 
 @pytest.mark.parametrize("schema", get_names(schemas), indirect=["schema"])
-@settings(**CORPUS_SETTINGS)
-@given(data=st.data())
-def test_corpus_negative(data, schema: Schema):
-    query = data.draw(from_schema(schema.raw, custom_scalars=schema.custom_scalars, mode=Mode.NEGATIVE))
+def test_corpus_negative(schema: Schema):
     parsed_schema = cached_build_schema(schema.raw)
-    query_ast = graphql.parse(query)
-    errors = graphql.validate(parsed_schema, query_ast)
-    assert errors, f"Query should be invalid in NEGATIVE mode: {query}"
+    strategy = from_schema(schema.raw, custom_scalars=schema.custom_scalars, mode=Mode.NEGATIVE)
+
+    @settings(**CORPUS_SETTINGS)
+    @given(query=strategy)
+    def inner(query):
+        query_ast = graphql.parse(query)
+        errors = graphql.validate(parsed_schema, query_ast)
+        assert errors, f"Query should be invalid in NEGATIVE mode: {query}"
+
+    inner()